Added map positioning, normal distribution, etc.
Added minimap.getLocalPosition(). This method will return the position
of the player, a point, or a texture on the map.

Added normalRandom(), normalRandomE(). This, like gauss, will return a
random number biased toward the median, using the central limit theorem.

Fixed glMapCoords() to not cache results if all coordinates are equal,
or the map was not found, or the function timed out.