Fixed mouse.rightClickOption() and other tweaks.
The mouse.rightClickOption() should no longer fail if the desired text
appears in the menu.

ogl.getTimeout() should now return the correct value.

Also, countDown.setTime() is now a function and not a procedure. It will
return the uInt32 passed to it - useful for resetting multiple timers
with the same value.