Added lodestone interface and mouse click methods.
Fixed a bug with ogl.getModels() failing, resulting in many empty
glModels being added to the glModelArray. Now, all models with an ID of
0 will be filtered from the array.

Added mouse.getClick(). This method, returning the string "red" or
"yellow", should be called directly after a mouse click, as it detects
the click textures.

Tweaked combat methods to be more consistent. Removed
combat.hasTarget(ansiString), which returned if the target name
contained ansiString. This was broken due to the combat mash-up and may
be readded in the future.

Added the lodestone interface. Users can now tell if the lodestone
interface is open, click a teleport by passing the name of the location,
and tell if the player is currently teleporting.