Added Debug Methods For OpenGL Types
Added ogl.getUpTime(). This returns a string in the format 0:00:00 which
can to be used for paints and progress reports.

Added debug methods for glChar, glCharArray, glModel, glModelArray,
glTexture, glTextureArray types. Each method has an optional parameter
which will clear the debug.

Added the smart.__graphics.drawPoly() method. This will trace a TPA.