All New Production, Added Some Documentation
Completely redid all production methods. These will now match those
recently changed in the actionBar interface. This is to improve
readability and prevent requiring an update, should the OpenGL types
change.

Added documentation to production. This is to prepare for the upcoming
ogLib documentation which is expected to closely match that of SRL-6.

Changed production.hasProduction(). This method will now return a
string. This will be one of: 'choose a tool', 'production', 'progress',
or ''.

Changed the production.clickClose() and production.clickMake(). These
are now production.setClose() and production.setMake() instead of
"click". This is because it will now randomly choose between using the
keyboard or the mouse.