All New Minimap, Fixed Dialogue, Etc
Added glChar.toPoint().

Added actionBar.getActionActive(). This method will return the active
state of an action. For example, when casting high alchemy, before the
item is clicked, high alchemy is active.

Rewrote and documented login.simba. This is a work in progress, as I
hope to fully support logins soon!

Rewrote and fully documented minimap.simba to be consistent with the new
coding style.

Changed minimap.getLocalPosition() and minimap.getScreebPosition()
methods to be much faster.

Added minimap.clickLocalPosition(). If the localPosition desired is not
visible on the minimap, it'll be ignored and return false!